refactor: Use CozyClient plugin mechanism for Sentry and Push
