+++ src/ducks/client/utils.js
@@ -1,3 +1,7 @@
+/**
+ * All this file should be moved to cozy-client
+ */
+
 export const isCollectionLoading = col => {
   if (!col) {
     console.warn('isCollectionLoading called on falsy value.') // eslint-disable-line no-console
@@ -30,4 +34,50 @@ export const associateDocuments = (
   }
 
   return originalDocument
+}
+
+export const isRevoked = async client => {
+  try {
+    await client.stackClient.fetchInformation()
+    return false
+  } catch (err) {
+    if (err.message && err.message.indexOf('Client not found') > -1) {
+      return true
+    } else {
+      return false
+    }
+  }
+}
+
+export const checkForRevocation = async client => {
+  const revoked = await isRevoked(client)
+  if (revoked) {
+    client.stackClient.unregister()
+    client.handleRevocationChange(true)
+  }
+}
+
+/**
+ * Updates the device notification token
+ *
+ * @param  {CozyClient} client
+ * @param  {String} token  - Device token, from Cordova Push Plugin
+ */
+export const updateNotificationToken = (client, token) => {
+  // Updates local and remote information
+  const clientInfos = client.stackClient.oauthOptions
+  client.stackClient.updateInformation({
+    ...clientInfos,
+    notificationDeviceToken: token
+  })
+}
+
+/**
+ * Get current device notification token
+ *
+ * @param  {CozyClient} client
+ * @param  {String} token  - Device token, from Cordova Push Plugin
+ */
+export const getNotificationToken = client => {
+  return client.stackClient.oauthOptions.notification_device_token
 }
